Implement project auto-detection using siteinfo (#42)
* Add project auto-detection.

This change implements project auto-detection by querying siteinfo.
If a project hasn't been specified explicitly via the -project command-line
flag or by providing a full hostname, now bmctool will query siteinfo's
projects.json to see if the node exists and which project is the right one.

* Fix tests.

To avoid having to mock httpGet everywhere, hostnames are always complete
with projectID so siteinfo is never queried except in the tests for
getProjectID itself.

* Update comments + remove flag mistakenly added again